requirements.txt
httpx==0.28.1
anthropic==0.49.0
beautifulsoup4==4.12.3
feedparser==6.0.11
reportlab==4.2.5
python-dotenv==1.0.1
rich==13.9.4
pdfminer.six==20231228
flask==3.1.3
google-api-python-client==2.169.0
google-auth==2.40.1
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
msal==1.31.1
# v2 web app
fastapi==0.115.12
sqlmodel==0.0.34
uvicorn==0.34.2
apscheduler==3.11.0
playwright==1.50.0
psycopg2-binary==2.9.10
authlib==1.3.2
itsdangerous==2.2.0
python-multipart==0.0.20
